Improvements for Sheets formatting and web preview
The Google Sheets auto-wrap now correctly avoids exceeding the width of the last line in a dialogue, which always has a prompt on it (regardless of line number - as it turns out, the text doesn't always scroll to a "page" interval at the end of a string).

The web preview now also reflects the scrolling behavior seen in the game better. Wahoo!